+++ flumine/order/process.py
@@ -71,11 +71,6 @@ def process_current_orders(
                     continue
             # process order status
             process_current_order(order, current_order, log_control)
-            # complete order if required
-            if order.complete:
-                market = markets.markets[order.market_id]
-                if order in market.blotter.live_orders:
-                    market.blotter.complete_order(order)
 
 
 def process_current_order(order: BaseOrder, current_order, log_control) -> None:
@@ -162,12 +157,6 @@ def process_betdaq_current_orders(
             continue
         # process order status
         process_betdaq_current_order(order, current_order)
-        # complete order if required
-        if order.complete:
-            for market in markets:
-                if order in market.blotter.live_orders:
-                    market.blotter.complete_order(order)
-                    break
 
 
 def process_betdaq_current_order(order: BaseOrder, current_order) -> None: